backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
pymongo==4.6.0
motor==3.3.2
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
streamlit==1.28.1
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
google-genai
requests==2.31.0
beautifulsoup4==4.12.2
